+++ examples/cct.rs
@@ -21,7 +21,8 @@ struct Args {
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let args = Args::parse();
-    let xyz: XYZ = XYZ::from_chromaticity(args.x, args.y, None, None)?;
+    let chromaticity = Chromaticity::new(args.x, args.y);
+    let xyz: XYZ = XYZ::from_chromaticity(chromaticity, None, None)?;
     let cct: CCT = xyz.try_into()?;
     let [t, d] = cct.into();
     let tint = d * 1000.0;