+++ src/spectrum.rs
@@ -628,6 +628,7 @@ mod tests {
             CIE1931
                 .xyz_from_spectrum(&red, None)
                 .chromaticity()
+                .to_array()
                 .as_ref(),
             &[0.64, 0.33].as_ref(),
             epsilon = 1E-5
@@ -644,12 +645,12 @@ mod tests {
     #[test]
     fn test_chromaticity() {
         let xyz0 = CIE1931.xyz_from_spectrum(&D65, None);
-        let [x0, y0] = xyz0.chromaticity();
+        let [x0, y0] = xyz0.chromaticity().to_array();
 
         let illuminance = D65.illuminance(&CIE1931);
         let d65 = D65.clone().set_illuminance(&CIE1931, 100.0);
         let xyz = CIE1931.xyz_from_spectrum(&d65, None);
-        let [x, y] = xyz.chromaticity();
+        let [x, y] = xyz.chromaticity().to_array();
 
         assert_ulps_eq!(x0, x);
         assert_ulps_eq!(y0, y);
@@ -702,24 +703,24 @@ mod tests {
 
     #[test]
     fn ee() {
-        let [x, y] = CIE1931
+        let chromaticity = CIE1931
             .xyz_from_spectrum(
                 &Illuminant::equal_energy().set_illuminance(&CIE1931, 100.0),
                 None,
             )
             .chromaticity();
-        assert_ulps_eq!(x, 0.333_3, epsilon = 5E-5);
-        assert_ulps_eq!(y, 0.333_3, epsilon = 5E-5);
+        assert_ulps_eq!(chromaticity.x(), 0.333_3, epsilon = 5E-5);
+        assert_ulps_eq!(chromaticity.y(), 0.333_3, epsilon = 5E-5);
     }
 
     #[test]
     fn d65() {
-        let [x, y] = CIE1931
+        let chromaticity = CIE1931
             .xyz_from_spectrum(&Illuminant::d65().set_illuminance(&CIE1931, 100.0), None)
             .chromaticity();
         // See table T3 CIE15:2004 (calculated with 5nm intervals, instead of 1nm, as used here)
-        assert_ulps_eq!(x, 0.312_72, epsilon = 5E-5);
-        assert_ulps_eq!(y, 0.329_03, epsilon = 5E-5);
+        assert_ulps_eq!(chromaticity.x(), 0.312_72, epsilon = 5E-5);
+        assert_ulps_eq!(chromaticity.y(), 0.329_03, epsilon = 5E-5);
     }
 
     #[test]
@@ -732,21 +733,21 @@ mod tests {
 
     #[test]
     fn d50() {
-        let [x, y] = CIE1931
+        let chromaticity = CIE1931
             .xyz_from_spectrum(&Illuminant::d50().set_illuminance(&CIE1931, 100.0), None)
             .chromaticity();
         // See table T3 CIE15:2004 (calculated with 5nm intervals, instead of 1nm, as used here)
-        assert_ulps_eq!(x, 0.345_67, epsilon = 5E-5);
-        assert_ulps_eq!(y, 0.358_51, epsilon = 5E-5);
+        assert_ulps_eq!(chromaticity.x(), 0.345_67, epsilon = 5E-5);
+        assert_ulps_eq!(chromaticity.y(), 0.358_51, epsilon = 5E-5);
     }
 
     #[cfg_attr(test, cfg(feature = "cie-illuminants"))]
     fn a() {
         let a: Illuminant = StdIlluminant::A.into();
-        let [x, y] = CIE1931.xyz_from_spectrum(&a, None).chromaticity();
+        let chromaticity = CIE1931.xyz_from_spectrum(&a, None).chromaticity();
         // See table T3 CIE15:2004 (calculated with 5nm intervals, instead of 1nm, as used here)
-        assert_ulps_eq!(x, 0.447_58, epsilon = 5E-5);
-        assert_ulps_eq!(y, 0.407_45, epsilon = 5E-5);
+        assert_ulps_eq!(chromaticity.x(), 0.447_58, epsilon = 5E-5);
+        assert_ulps_eq!(chromaticity.y(), 0.407_45, epsilon = 5E-5);
     }
 
     #[test]